Create README.md
This message is for most of us. The ones who are not as bold as we want to be. Myself I'm open and laid back until you met or talk to me, then you find out I have a pulse and not a square. Dig it or not! WHat's it too Ya! Villainz! 

VvMichaeals